+++ jdaviz/configs/imviz/plugins/aper_phot_simple/aper_phot_simple.py
@@ -76,6 +76,8 @@ class SimpleAperturePhotometry(PluginTemplateMixin, ApertureSubsetSelectMixin,
         ``multiselect=False``
     * ``table`` (:class:`~jdaviz.core.template_mixin.Table`):
         Table with photometry results.
+    * ``cube_slice``
+      Current slice wavelength being used for aperture photometry (cubes only, read-only).
     """
     template_file = __file__, "aper_phot_simple.vue"
     uses_active_status = Bool(True).tag(sync=True)
@@ -196,7 +198,10 @@ class SimpleAperturePhotometry(PluginTemplateMixin, ApertureSubsetSelectMixin,
                   'export_table', 'fitted_models', 'current_plot_type',
                   'fit_radial_profile', 'plot')
 
-        return PluginUserApi(self, expose=expose)
+        if self.config == 'Imviz':
+            return PluginUserApi(self, expose=expose)
+        else:
+            return PluginUserApi(self, expose=expose, readonly=('cube_slice',))
 
     @property
     def fitted_models(self):
@@ -509,7 +514,7 @@ class SimpleAperturePhotometry(PluginTemplateMixin, ApertureSubsetSelectMixin,
     @property
     def _cube_slice_ind(self):
         # TODO: performance improvements, change to listen to slice change event
-        slice_plugin = self.app._jdaviz_helper.plugins.get('Slice', None)
+        slice_plugin = self.app._jdaviz_helper.plugins.get('Spectral Slice', None)
         if slice_plugin is None:
             return None