+++ jdaviz/configs/imviz/plugins/aper_phot_simple/aper_phot_simple.vue
@@ -27,13 +27,14 @@
       :api_hints_enabled="api_hints_enabled"
     />
 
-    <div v-if='config == "cubeviz" && is_cube'>
+    <div v-if='is_cube'>
       <v-row class="row-no-outside-padding row-min-bottom-padding">
         <v-col>
           <v-text-field
             :value="cube_slice"
             class="mt-0 pt-0"
-            label="Slice wavelength"
+            :label="api_hints_enabled ? 'plg.cube_slice' : 'Slice wavelength'"
+            :class="api_hints_enabled ? 'api-hint' : null"
             hint="Extracts photometry from currently selected cube slice"
             disabled
           ></v-text-field>