+++ jdaviz/configs/specviz2d/plugins/cross_dispersion_profile/cross_dispersion_profile.vue
@@ -1,6 +1,7 @@
 <template>
   <j-tray-plugin
     plugin_key="Cross Dispersion Profile"
+    :api_hints_enabled.sync="api_hints_enabled"
     :description="docs_description"
     :uses_active_status="uses_active_status"
     :keep_active.sync="keep_active"
@@ -9,7 +10,7 @@
     :disabled_msg="disabled_msg"
     :scroll_to.sync="scroll_to">
 
-    <v-row v-if="plot_available" style="padding: 0px"
+    <v-row v-if="plot_available" style="padding: 0px">
       <jupyter-widget :widget="plot_widget"/>
     </v-row>
 
@@ -18,10 +19,15 @@
       :selected.sync="dataset_selected"
       :show_if_single_entry="false"
       label="Data"
+      api_hint="plg.dataset ="
+      :api_hints_enabled="api_hints_enabled"
       hint="Select the data to compute the profile."
     />
 
     <v-row>
+      <span v-if="api_hints_enabled" class="api-hint" style="font-size: 10px">
+        plg.pixel =
+      </span>
       <v-col cols="9" style="padding: 0px">
         <v-slider
           :min="0"
@@ -31,6 +37,7 @@
           hint="Pixel on spectral axis."
           persistent-hint
         >
+        </v-slider>
       </v-col>
 
       <v-col cols="3" style="padding: 7px">
@@ -49,11 +56,16 @@
     <plugin-switch
       :value.sync="use_full_width"
       label="Use full cross-dispersion width."
+      api_hint="plg.use_full_width = "
+      :api_hints_enabled="api_hints_enabled"
       style="font-size: 10px;"
     />
 
     <div v-if="!use_full_width">
       <v-row>
+        <span v-if="api_hints_enabled" class="api-hint" style="font-size: 10px">
+          plg.y_pixel =
+        </span>
         <v-col cols="9" style="padding: 0px">
           <v-slider
             :min="0"
@@ -63,6 +75,7 @@
             hint="Center of profile on cross-dispersion axis."
             persistent-hint
           >
+          </v-slider>
         </v-col>
 
         <v-col cols="3" style="padding: 7px">
@@ -79,6 +92,9 @@
     </v-row>
 
       <v-row>
+        <span v-if="api_hints_enabled" class="api-hint" style="font-size: 10px">
+          plg.width =
+        </span>
         <v-col cols="9" style="padding: 0px">
           <v-slider
             :min="0"
@@ -88,6 +104,7 @@
             hint="Width in cross-dispersion axis, in pixels."
             persistent-hint
           >
+          </v-slider>
         </v-col>
 
         <v-col cols="3" style="padding: 7px">
@@ -102,8 +119,6 @@
           />
         </v-col>
       </v-row>
-    </div>
-
     </div>
 
   </j-tray-plugin>