+++ jdaviz/core/template_mixin.py
@@ -2708,6 +2708,7 @@ class SubsetSelect(SelectPluginComponent):
 
         # Use Glue's to_mask method directly rather than translating the whole Data object to
         # output class (e.g., Spectrum)
+        glue_mask = None
         for sg in self.app.data_collection.subset_groups:
             if sg.label == subset:
                 if hasattr(self.plugin, "cube_fit") and self.plugin.cube_fit:
@@ -2717,6 +2718,11 @@ class SubsetSelect(SelectPluginComponent):
                     data = self.app.data_collection[dataset]
                 glue_mask = ~sg.subset_state.to_mask(data)
                 break
+
+        # If no subset group was found return True (no masking)
+        if glue_mask is None:
+            data = self.app.data_collection[dataset]
+            glue_mask = np.zeros(data.shape, dtype=bool)
 
         if self.app.data_collection[dataset].ndim == 3 and glue_mask.ndim == 1:
             data = self.app.data_collection[dataset]
@@ -4824,7 +4830,8 @@ class AddResults(BasePluginComponent):
         if self.app.config in CONFIGS_WITH_LOADERS and format is not None:
             self.app._jdaviz_helper.load(data_item,
                                          loader='object', format=format,
-                                         data_label=label, viewer=[], **load_kwargs)
+                                         data_label=label, viewer=load_kwargs.pop('viewer', []),
+                                         **load_kwargs)
         else:
             # NOTE: eventually remove this entirely once all plugins are set to go through
             # the new loaders infrastructure above